Add Blaze of Light (#22)
