+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@focused-will/components": "^0.3.5",
+    "@focused-will/components": "^0.3.6",
     "@heroicons/react": "^1.0.5",
     "@radix-ui/react-toast": "^1.1.1",
     "@testing-library/jest-dom": "^5.11.4",
